Write VISA session
